Add files via upload
Objective

In this project, we will be able to create a countdown timer using Python. 
The code will take input from the user regarding the length of the countdown in seconds. 
After that, a countdown will begin on the screen of the format ‘minutes: seconds. We will use the time module here.

Instructions
In this project, we will be using the time module and its sleep() function. Follow the below steps to create a countdown timer:

Import the time module.
Ask the user to input the length of the countdown in seconds.
This value is sent as a parameter ‘t’ to the user-defined function countdown(). Any variable read using the input function is a string. So, convert this parameter to ‘int’ as it is of string type.
In this function, a while loop runs until time becomes 0.
Use divmod() to calculate the number of minutes and seconds. You can read more about it here.
Now print the minutes and seconds on the screen using the variable timeformat.
Using end = ‘\r’ we force the cursor to go back to the start of the screen (carriage return) so that the next line printed will overwrite the previous one.
The time.sleep() is used to make the code wait for one sec.
Now decrement time so that the while loop can converge.
After the completion of the loop, we will print “Fire in the hole” to signify the end of the countdown.